notebook_app_by_cc/backend/requirements.txt
django==4.2.30
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
